Added sticky navbar, new homepage styling, global variables
What's new:
Small changes to navbar so it always stays on top and it sticks to top of the page.
Homepage styles got changed.
Global variables got added.

To fix:
Icons are to top left of input box, not inside of the box.
Placeholder colours are slightly different.
Job card hover doesn't work.
Size for smaller screens isn't optimized anymore.