Update HomeScreen Search Bar
What's new:
The previous commit didn't include all changes to the search bar.

To fix:
(The same as the previous commit)
Icons are to top left of input box, not inside of the box.
"Dodaj ofertę pracy" in the footer isn't a link yet.
Job card hover doesn't work.
Size for smaller screens still isn't optimized.